Updated the stock price's data type to numeric/decimal instead of integer

+++ db/migrate/20210312013145_change_data_type_to_sell_order.rb
@@ -3,6 +3,6 @@ class ChangeDataTypeToSellOrder < ActiveRecord::Migration[6.0]
     change_column :sell_orders, :price, :numeric, precision: 8, scale: 2
   end
   def down
-    change_column :sell_orders, :price, :bigint
+    change_column :sell_orders, :price, :integer
   end
 end
